- Added spear chick. - Shuffled units. - Added second human map. - Tweaked battle mechanics. - Fixed the editor. - Added single point brush.
git-svn-id: http://localhost/svn/warlords@354 f82a8f4f-13e0-4542-855c-6df575116ada

+++ Source/Map.h
@@ -94,7 +94,7 @@ public:
 	void DrawDebug();
 
 	void GetMapSize(int *pWidth, int *pHeight) { if(pWidth) *pWidth = mapWidth; if(pHeight) *pHeight = mapHeight; }
-	void GetCursor(float x, float y, int *pX, int *pY);
+	void GetCursor(float x, float y, float *pX, float *pY);
 	void GetVisibleTileSize(float *pWidth, float *pHeight);
 
 	void SetOffset(float x, float y);
